+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## [Unreleased]
 ### Changed
 - Ampache: A few more actions now support pagination with offset and limit: `artist_albums`, `artist_songs`, `album_songs`, `search_songs`
+- Subsonic: Added support to `getArtistInfo` to identify the artist using a track, an album, or a folder ID
+  [#906](https://github.com/owncloud/music/issues/906)
 
 ### Fixed
 - A performance problem affecting Subsonic method `getArtist`, Ampache action `artist_albums`, and a few other functions